A piece of assistance for reading Go assembler.
I regret to admit this has existed since last Sept 8th -- or perhaps
Aug 4th?  who can say, really -- according to filesystem metadata, and
I'm only getting around to committing it now.

The usage is typically like:

	gof build -gcflags "-S" ./... 2>&1 |
	go run go-assist/main.go '*TypeName).MethodName'

This just selects the body of assembly for that function.  (Doing so
with line-oriented grep is not possible, because information is carried
by the indentation, which requires more power to recognize.  Perhaps
there are other shellish incantations that could do it, but I found
banging out this quick helper to be easier than puzzling over it.)

Note that this particular example is for methods on a pointer; the
parens and star are absent if it's a method on a bare struct.

Signed-off-by: Eric Myhre <hash@exultant.us>